project: add `_get_project_directory_hash` method
Due to namespace collisions when working on the same project in
multiple directories simultaneously, provide an interface to
consistently determine which "instance" of the project we are
working from - notable the current working directory.

Future use could include providers incorporating this hash into
their naming schemes.  At minimum, remote-build will use it in
the following commits.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>