remote-build: introduce LaunchpadGitPushError
When an error occurs during git pull, the user is presented
with a "pull" exception due to use of the git sources module.
Currently all command failures for sources are assumed to be
pull errors, so we just catch this error and raise a more
appropriate one for the user.

Add a test case for git push failures to validate.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>