meta: force grade devel when using non stable bases
Check the channel used to install the base snap and set the grade of the snap
build accordingly with appropriate warnings.

If grade default setting has been moved to Snap.to_dict to avoid warning
duplicates from the various locations a Snap object is instantiated from.
This, for now, also makes applying a default that is only relevant to the
final result a property of the transformation and not the created object.

Snap.validate has been extended to check for conflicting settings with
regards to the expected grade and the one found in snapcraft.yaml (if any)
and error if the action cannot be satisfied.

Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>