extensions: skip icon cache creation for theme and runtime snaps
Some icon themes, such as cursor-only themes, can't have a cache.
However, desktop-exports still tries to create a cache for
every icon theme, which is surprisingly expensive. It takes up to
30% of the time of a cold desktop-launch run. This is partly due
to the script creating additional directories and softlinks
during this process.

This commit skips icon cache creation for everything in the theme
and runtime snaps because those should already contain caches.

With this commit, the cold start time of the desktop-launch script
is reduced by ~30%.