remote-build: explicitly default build arch to host arch
If architectures is not defined in project, and --arch is not
used, default to building the host architecture.

Without this, it appears that Launchpad will build with a default
set of `i386` and `amd64`.  Better to define this architecture
explicitly and matching the standard snapcraft behavior seems
like an appropriate choice.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>