lifecycle: store the required grade depending on the base
Query the store if the required grade depending on the base has not been stored.
If not, query the Snap Store return:
- stable if the base is available on the stable channel
- devel if the base is not found on the stable channel

Also return devel if the base is nowhere to be found.

Store the result in GlobalState's set_required_grade.

Add a global fake to the entire unit test suite to avoid calling to the Snap
Store when this code path is traversed.

Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>