meta: ensure Snap's `assumes` is initialized as a set
When reading the config data dictionary, convert the list to a set.

The `assumes` property is defined as a set in Snap meta, but derives
from a list in the the snap meta.  This causes an exception when
using `command-chain`, but `command-chain` is not already in
the `assumes` (due to the use of Snap's usage of `.add()` on the
set, which is not available for lists).

Add a test for coverage/verification.

Fixes SNAPCRAFT-16C

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>