cli: exempt containers from requiring --destructive-mode
This is not a change in behavior, but this codifies
the rule explicitly rather than relying on the order of
provider resolution.  The ordering was changed in a previous
commit and caused a test failure for `test_host_container()`
in tests/unit/commands/test_build_providers.py.

This fixes that test and makes it clear what the expected
behavior should be in `_sanity_check_build_provider_flags()`.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>