project loader: explicitly check match in ToStatement grammar processing
A match may be None in `_extract_to_clause_selectors()`.  Check for
None and raise ToStatementSyntaxError() explicitly.  This should
be functionally equivalent because the later AttributeError exception
would be thrown and caught when calling `.group()` on None.  Remove
the AttributeError catch as it should no longer be necessary.

Other than better form, this resovles a mypy error on uprev.

Additionally add missing Optional annotation for `body`.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>