project loader: explicitly check match in OnStatement grammar processing
A match may be None in `_extract_on_clause_selectors()`.  Check for
None and raise OnStatementSyntaxError() explicitly.  This should
be functionally equivalent because the later AttributeError exception
would be thrown and caught when calling `.group()` on None. Remove
the catch for AttributeError as this no longer should happen.

Other than better form, this resovles a mypy error on uprev.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>