lxd provider: address mypy uprev errrors
Move __init__() up to top of instance class methods, after class methods.
This is due to mypy defining the type as it sees it, and so _container
was already defined before init.

For each method using _container, check if it is None and raise
a RuntimeError() if it occurs.  It shouldn't happen and is almost
certainly a developer error.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>