store: validate type for snap_ids in _human_readable_acls()
The return type is a Union[List[str], str, None].  None is effectively
checked for already, but string is not.  If it were a string,
the `for snap_id in snap_ids` loop would incorrectly iterate over
the characters in the string.

Add a check to ensure it is an instance of list, and raise
a RuntimeError if not (it shouldn't happen).

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>