meta: detail Snap's to_dict()
Similar to that done for from_dict(), get rid of __dict__ usage
in to_dict() just handle each field directly.  More verbose,
but detailed.

Shouldn't affect behavior, but minor changes include the use of
OrderedDict() over dict() and a simpler `if <field>` rather than
checking for None.  It's not valid for these fields to be empty
strings.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>