meta: detail Snap's to_dict()
Similar to that done for from_dict(), get rid of __dict__ usage
in to_dict() just handle each field directly.  More verbose,
but detailed and more flexible.

Shouldn't affect behavior, but this changes the use of dict() to
OrderedDict() to retain ordering for `apps`, `hooks`, `plugs`,
and `slots`.  Their keys were being sorted, but likely lost without
the the use of OrderedDict().

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>