xattrs: ignore errors if SNAPCRAFT_BUILD_INFO is unset
As SNAPCRAFT_BUILD_INFO is not always enabled, ignore errors
setting xattrs to record origin stage packages.  This data
will not be used if there is no manifest being generated.

This should prevent unnecessary errors being raised to the user
if their file system doesn't support xattrs (or are otherwise
limited from reading/modifying them).

Introduce XAttributeError() to improve the error messaging
to the user should reading/writing xattrs fail.  Mark as
reportable for now, but ultimately this should be False once
the feature is stable and extensively used by the community.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>