build providers: only set the snapd flag when installing snapd
When setting the experimental.snapd-snap to true, the system goes into a
deadlock expecting the snapd snap to be there. Given that the snapd snap
is not install when core is injected it causes a deadlock on the build
environment. Avoid the situation by only setting this flag when the snapd
snap is to be installed.

Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>