dirs: support --user install on Linux
Snapcraft includes explicit support for being installed in a number of
situations, from the Debian package, to the snap, to a virtual
environment. It supports a number of strategies on Windows as well,
including being installed into the user base. There are times when
snapcraft needs to be installed into the user base on Linux, too, for
example when hacking on plugins related to Python 2, such as Catkin. Add
support for running out of the user base on Linux.

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>