common: generate run scripts which can execute independently
To support debugging, update run() to create a script that accounts
for the parameters `cwd` and `env` passed along to check_call(),
stripping those parameters (if present).

No longer remove script generated, but instead persist it in
/tmp/snapcraft-<pid>/run-<#>.sh for optional debugging.

Now a developer can look at the script and re-run it independently of
snapcraft for debugging/development purposes.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>