remote-build: configurable timeout/deadline for builds
Introduce --launchpad-timeout option, defaulting to disabled.
Reword ProviderNotReadyError -> RemoteBuildTimeoutError.

Honor timeout deadline for both start_build() and monitor_build() loops.

This was done because Launchpad was taking an unusally long time
to accept a build request, and the timeout prevented the user from
addressing it.  Instead, disable the timeout by default, and only
deadline if the user insists upon it.

Update tests accordingly.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>