errors: introduce SnapcraftPluginAssertionError()
If a plugin detects an error prior to attempting to build a part's
source, it can raise SnapcraftPluginAssertionError() with a brief
reason and optional details.

SnapcraftEnvironmentError() was also a viable choice, but it's slated
to be nuked and as it is too generic.  This seems like a fair middle
ground that can be re-used across the the plugins when they make an
assertion that fails to check out at runtime.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>