wstool: don't rely on host git
The Catkin plugin utilizes wstool to support rosinstall files. While
wstool is fetched into its own install area so as to not contaminate
either the host or the snap, it doesn't properly use the git that is
fetched along with it. This means that merging rosinstall files that
include git repositories only works if git is already installed on the
host. If the host doesn't include git, it errors out.

Properly set the variables necessary to run git from wstool's install
area, thereby no longer relying on the host already having git
installed.

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>