meta: remove Application's `prepend_command_chain`
Remove `prepend_command_chain` member - it's ambigiuous by name and
makes more sense to just modify the command-chain directly.

Rename and simplify snap packaging's `_generate_command_chain` ->
`_generate_snapcraft_runner` as it really just creates the runner,
if required.

This indirectly fixes _ensure_command_chain assumption, because
it was not checking for "prepend".  So the expected snap.yaml
for the adapter test is updated with the proper assumes that's now
being set.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>