rust: add support for workspaces (#2842)
If a plugin detects an error prior to attempting to build a part's
source, it can raise SnapcraftPluginAssertionError() with a brief
reason and optional details.

SnapcraftEnvironmentError() was also a viable choice, but it's slated
to be nuked and as it is too generic.  This seems like a fair middle
ground that can be re-used across the the plugins when they make an
assertion that fails to check out at runtime.

Unfortunately, `cargo install` does not support workspaces yet:
https://github.com/rust-lang/cargo/issues/7599

Alternatively, we use `cargo build`, which will respect the Cargo.lock
configuration.  It however, does require a bit more hoop jumping
to determine which binaries were built and install them.

Introudce _install_workspace_artifacts() to install the built
executables into the correct paths.  Testing has covered executables
and libraries, though dynamic linking is not quite yet supported
by the rust plugin (at least in my testing, it will have unmnet
dependencies on libstd-<id>.so).  We can address that feature gap
in the future, but likely doesn't affect snap users because they
are probably using the standard linking process which doesn't require
libstd (likely due to static linking of those dependencies).

`cargo build` has an unstable flag option for `--out-dir` which
may simplifiy the install process, but is currently unavailable
for stable use:
https://doc.rust-lang.org/cargo/reference/unstable.html#out-dir

Add/update unit tests for coverage and add a spead test.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>