elf: extract build ID and presence of debug info (#2229)
Updates the ElfFile class to extract the build ID and presence of debug info in a file. This is needed to identify candidates to have their debug info separated, and to identify the debug info corresponding to installed ELF files.

Also took the opportunity to simplify ElfFile._extract to remove dependence on section names where they aren't actually relied on. For example, while the ELF interpreter is usually found in the .interp section, the kernel actually looks for it in the PT_INTERP segment. You could store the name in a differently named section of that segment and it would still work.