snaps: account for forwarded (effective) installation channels
Snapcraft assumes a channel is invalid if it doesn't find it
in the list provided by the /v2/find snapd API. However, snapd
may instead be forwarding an installation request to a safer-risk
channel.  Account for this by determining the effective snap's
channel, and use this for validating the snap in question.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>