meta: handle plug & slot string objects
Although undocumented, the plug and slot objects may in
fact be a string, such as the case here:
https://git.launchpad.net/~snappy-hwe-team/snappy-hwe-snaps/+git/modem-manager/tree/snapcraft.yaml#n16

If the object is a string, use that string as the interface.

To ease the handling of this, we introduce cls.from_object() for
Slot and Plug to handle None, string, and dict cases. Snap will
now use these methods when populating the slots and plugs, instead
of cls.from_dict().

Add tests for coverage of from_object() and the string case.

LP #1859806

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>