extensions: fix merging of build-environment
When merging "build-environment", merge_lists will fail
because OrderedDict is unhashable when calculating the "seen"
set.

In addition to this failure, build-environment should be merged
with the extension definitions being listed first, followed by
the user definitions.  So we append the lists in that order.

Add tests to show coverage for "build-environment" using the
part_snippet.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>