plugs/slots: match output format read in snapcraft.yaml
Snapcraft recently started always writing the long-form
of plugs for consistency, where the `interface` attribute
is always defined.

Unfortunately this seems to have tripped up a bug in review-tools
that won't likely be fixed for a while.  This commit restores
the 3.8 behavior to ensure the output snap.yaml matches the syntax
used in the snapcraft.yaml.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>