lifecycle: only warn when a default provider snap is missing
Snapcraft gained the ability to check for artifacts from snaps listed as plugs
when using the content interface.

The side effect, is that the snap needs to be published on the main Snap Store,
this is a requirement brand stores cannot meet, so instead of failing,
Snapcraft resorts to warning instead.

Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>