elf: read ELF type when extracting attributes
The elf type may be used for decision making when handling
executables.  This will be useful for determining what to do
when stripping executables, as different types should be handled
differently.

Use the enum strings provided by pyelftools instead of integer
value.  Initialize to "ET_NONE", but always read the entry
when extracting attributes from file.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>