lifecycle: raise detailed error if mksquashfs fails
Sentry is full of bug reports related to RuntimeError for failing
to execute mksquashfs.  However, there is no clue as to the failure,
because snapcraft just prints it to logger.  Instead of logging the
error, include the output as part of the raised RuntimeError.

If we can get a better idea of the nature of the failures, we can
further refine handling here.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>