python plugin: do not leak snapcraft's site-packages
The current sitecustomize generated by the python plugin always considers
adding site-packages from $SNAP which is only meant for runtime. This is a
problem as snapcraft runs in the context of a snap, so $SNAP always exists
in the environment.

When building, snapcraft exports SNAPCRAFT_INTERPRETER, use this as a condition
to export site-packages found under $SNAP.

LP: #1860884

Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>