meta: do not prime commands with adapter == "none" (#2912)
Snapcraft is priming (modifying) the commands when adapter is set
to "none".  These commands should be passed through as-is and never
modify them.  This will be consistent with snapcraft's behavior with
regards to wrappers and snapcraft-runner (it does not generate wrappers
or modify the command-chain).

Add unit test for "none" case and update adapter spread test to
include a case for a adapter "none" with a python script command.

LP #1862163

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>