plugin handler: process elf files only if base is specified (#2926)
ELF files are always processed when `type: app`, where `app`
is the default type.

This causes an issue when snapd is building without specifying
the type, as it gets treated as `type: app` and its files
are processed like typical app snaps.  The schema currently
allows for this because it fails to set the default type `app`
in time for validation.

Since snapd does not have a base, check for the existence of the
base before attempting to process ELF files.  This will prevent
the error reported in LP #1857019 and allow snapd to build without
a type specified.

Once snapd moves to `type: snapd`, we can fix the schema to
ensure that the invalid configuration that led to this error
does not pass verification.

Add basic snapd-like spread test `snapd-workaround` to ensure that
snapd is buildable and has the expected snap.yaml.  Once snapd
migrates to `snapd` type, this test will be migrated.

LP #1857019
LP #1862642

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>

Co-authored-by: Sergio Schvezov <sergio.schvezov@canonical.com>