build providers: clean up LXD startup message
This message about waiting for cloud-init is an implementation detail.
This has confused users, some thought that their data was being
_sent to the cloud_ which is not what is happening considering
what cloud-init is meant to be doing.

The code base is also moving away from cloud-init to profile containers
for Snapcraft, it might as well be ready for that moment.

Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>