snap-packaging: remove broken host-compatibility check for runner (#2951)
The check is incorrect as-is because is_host_compatible points to a
function, not a bool, effectively evaluating to true (always).

Just remove the check and generate the environment as is currently
being done anyways.  Even if the host were incompatible, the
environment will be more correct that omitting it.  Building on
an incompatible host is user-beware and this is one of the risks.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>