unit tests: move to pytest
All tests have been migrated to work with pytest.
This means that testscenarios as such is not longer being used for unit
tests and using a shim that parametrizes tests for pytest.

For the cases where the unit.TestCase was to ingrained in the test, the
strategy to replace scenarios has been to go with test duplication. This
has the benefit that testscenarios removed where it is now possible to
run an individual test.

pytest specific fixtures were added to better support some test cases.
Most tests are unchanged, but would benefit from smaller fixtures as
those added by pytest for faster test setup.

runtests.sh has been adapted to use pytest with pytest-cov, but keeping
unittest as the runner for integration tests.

The spread tests were also updated, but it would be good to see them not
run in spread anymore.

Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>