pluginhandler: do not search installdir or stagedir for dependencies (#2942)
The search paths already include primedir which has all of the primed
and staged files at this point.  The installdir and stagedir may
contain files that are not shipped in the final snap, so limit search
to primedir as it should be fully populated.

- Replace _handle_dependencies() with a simpler _warn_missing_dependencies()
and _calculate_dependency_paths() called from _handle_elf().

- Add a couple useful debug prints.

- Add two tests to cover the case where the dependency is found, and
another where the dependency is missing.  For the latter, the lib is
removed using `stage` configuration.

This last test would have previously failed, because the dependency
was found in installdir, and ignored by the missing dependency resolver.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>
Co-authored-by: Sergio Schvezov <sergio.schvezov@canonical.com>