multipass: refactor copy_files into push_ and pull_file using stdio
Strictly confined multipass won't be able to read the files, the only reliable way is to pass the data through stdio.

Unfortunately due to: https://bugs.launchpad.net/ubuntu/+source/snapd/+bug/1849753 we can't simply rely on std{in,out}=open(...).

The only memory-wise way to achieve this is to read/write the files in chunks.