repo: ubuntu implementation refactoring with initial package-management
Instead of allowing support for multiple apt instances, we consolidate
to the system's configuration.  As build-environments are instianted for
builds, we do not have the prior concerns of impacting the user's system
(in destructive mode).  Even in destructive mode, we shouldn't be
causing damage.

- Use a single instance of the apt cache, which should improve performance
due to the penalty of recreating caches in numerous points.  It also
simplifies usage without the requirement of copying/sharing instances
for what is (now) a system-wide resource.

- Remove old scheme for gathering stage packages, and simply do a scan
ourselves for the packages and their dependencies.  Then fetch them
using python-apt's fetch_binary().

- Rework of the ROS plugins to remove their current usage of Ubuntu()
and replace it with a simpler set of APIs to simply install the required
repository and keyrings.

TODO:
- squelch python-apt's debug log
- fixup tests

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>