repo: always use host source lists and remove those found in plugins
The Ubuntu object is now responsible for ensuring that all of
the configured apt sources are merged (performed in
AptCache._collected_sources_list()).

- (Re)define plugin.PLUGIN_STAGE_SOURCES to be a list of strings of
repositories to be added.  The base plugin will return an empty
list if the plugin doesn't define it.  There was a bit of mixed
usage of str vs. list, so we make it explicit.

- Remove the redundant sources specified by the catkin and colcon
plugins, we do not support building for non-native releases.

- Update tests accordingly, removing catkin's tests for verifying
sources are formatted for cross-release build (i.e. base does
not match build environment), which is no longer supported.

- Remove unused reference to plugin._PLUGIN_STAGE_SOURCES.

- Remove unused reference to plugin._PLUGIN_STAGE_KEYRINGS.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>