plugins: introduce v2.PluginV2 and v2.NilPlugin
There's just enough scaffolding in place to allow for the plugin to
load.

PluginHandler has been extended to determine when a PluginV2 is being
used (PluginHandler should be renamed to PartHandler).

Some extra attribute were added to PluginV2 to minimize the changes
required in PluginHandler.

Old code handling state file migration in PluginHandler._do_build was
removed as that was done before the migration to bases (meaning that
no base using Snapcraft project needs this).

Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>