remote build: remove artifact sanity check (#3021)
A snapcraft.yaml may have a configuration of `build-on` that allows
Launchpad to pick from a set of architectures that will run on
multiple architectures.

This will fail our sanity check which assumes the arch, which may
in fact end up as `multi`.  There isn't really any need to sanity
check the artifacts anyhow, so we'll remove this check and download
whatever Launchpad provides.

This exposed a problem with a unit test that was missing the fact
that a download was previously skipped due to sanity check. Fixed.

LP: #1870730

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>