tests: speed up step, pack and clean command unit tests
The unit tests are not testing the actual command logic with actual
assertions on methods used with --destructive-mode, --use-lxd and with
the default provider (multipass).

Test the clean command implementation and not its dependencies.
This takes this suite from ~90s to ~0.5s.

The step related commands dropped from around ~15s-40s to ~0.5s.

Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>