build providers: wait for systemd and better nameserver setup on LXD
systemd has a sub-command "is-system-running", we care about two
results:

- running: which is implied with exit code 0
- degraded: which exits with code 1, for which we need to parse the
            output.

The reason to parse for degraded is that LXD containers usually work
with systemd in a degraded state.

Additionally enable the build-provider test suite for core20, also
using the common make-hello instead of the in "suite" one (symlinked).

when systemd-resolved is running on the system, it will write the
actual resolv.conf in a prescribed location, which is now picked up as
the default to use the best local-to-host nameserver. The use of the
environment variable still overrides any setting for the cases where
it may be needed.

Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>