repo: restore marked-install strategy for apt-cache
Re-introduce refactored AptCache to wrap apt-cache usage.
Use it in read-only mode for working with the host cache
for use with build-packages, and use it to create a new
transient cache for use with handling stage packages. It
will symlink the host /etc/apt configuration so it matches
the host configuration, but allow for an empty cache to mark
packages and dependencies for fetching.

The strategy of marking stage packages ourselves was more
complicated than it appeared, and this achieves a far better
result with less risk of affecting existing snaps.

Additionally this fixes primed-stage-packages, which mistakenly
was using a `:` instead of `=` to split the name and version.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>