package repositories: avoid setting up on host when not the target
The keys were loaded and setup during the project_loader's init.
This is an issue as this is loaded before launching the build
environment which causes host interactions before launching a build
environment.

Defer the key setup and build_packages/build_snaps collection until
needed and call from the lifecycle.

Tests were updated to be segregated to each concern. The Manifest
generation keys which started to fail due to changes required were
completely removed from the lifecycle tests as they are not of concern
to the scope of lifecycle (and are covered by integration tests which
these behaved like).

Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>