repo: use persistent cache for fetching stage packages
It several seconds for the apt cache to to be updated and several
more to re-verify the hashes of archives that were already fetched.
It may even take minutes depending on performance and how many
stage-packages are being requested.

Set results to expire in 24 hours as the apt packages should be
refreshed periodically, particularly staged packages.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>