pcache: introduce persistent cache decorator
Cache results of wrapped function, returning the last known-good result,
if available.  Configurable expiration times may be set, with a default
of one-day.

The recorded items have an identity key derived from:
- function name
- function arguments
- function keyword arguments

All arguments to the wrapped function must be representable via repr().
The arguments must be represented in an uniquely identifiable and
repeatable manner to ensure accuracy and prevent collisions.

Note that this current implementation is not optimized for performance.
The database is opened and closed on-demand which carries a cost.  It could
be updated to keep the database open across the lifetime of the process,
but that may complicate usage with multiple processes (e.g. snapcraftctl).
For our current use cases, this is the safe and appropriate default.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>