elf: fix parsing of notes after patchelf mangling
We cannot rely on iterating on the notes via the note segment.
Patchelf may move note sections into other segments, and worse
yet, possibly leave the note segment standing with invalid data.
Here we instead iterate over each section and (more) reliably
gather the note information.

There is some improvement that should be done on the patchelf
side, as `readelf` complains on these binaries, and other issues
have surfaced in the past (e.g. mesa in classic snaps).  However,
this will hopefully address snapcraft build-time issues with not
only our patched files for classic snaps, but externally patched
binaries that are being packaged.

Add a couple of developer logs to indicate when we are parsing
an ELF, so future investigations are easier when something explodes.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>