go v1 plugin: go.mod support for 1.11 and 1.12
Add limited support for using Go Modules for 1.11 and 1.12 of Go.

The limitation when using 1.11 and 1.12 is that only the main module is
installed. This is due to limitations in "go build" in those versions,
to find the correct target, a new method "_get_module" is introduced.

A warning is issued when using 1.11 and 1.12 with a link on how to
enable support for Go Modules with precautions to only print the warning
once per execution.

Some general cleanup in GoPluing._build took place to remove repetition
of "domain" specific flag settings for when building and relinking.

LP: #1879315

Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>