cli: migrate close to use the new channel map
The implementation was moved to the CLI as being done for all commands.

With this, the new implementation for displaying a channel map is used,
which avoids the bug in the channel_map_tree returned in close that is
not progressive release aware and shows channels mistakenly as open.

The closed_channels API results are skipped as well as they are not
really track aware, returning mixed results for tracks in some scenarios
and plain risks for others (when latest is used, which is not allowed
when using more than one track).

The tests have been simplified to the concerns of the close CLI results
as display variants of the channel map tree are already done under test_status.

A missing todo checklist entry was added to take care of the upload logic.

Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>