plugin handler: fix local plugin class detection with direct imports
If a local plugin imports a snapcraft in-tree plugin directly,
e.g. `from snapcraft.plugins.v1 import AutotoolsPlugin`,
_get_local_plugin_class() will find an return the AutotoolsPlugin
attribute, not the desired plugin class, which inherits from
AutotoolsPlugin.

This change filters out any attributes that has a __module__ from
snapcraft.plugins.v1.  This allows us to remove the direct check for
PluginV1, which was likely the first case of filtering for this problem.

Spread tests:
- add another x-local test with this logic for PluginV1.
- modify from-nilplugin to use the direct import of NilPlugin.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>