cli: use snap pack instead of mksquashfs
Moving to using snap pack for everything as agreed with the snapd team.
This makes it easy to ensure snapcraft is not deviating from snapd
standards. Additionally, --check-skeleton is no longer needed as snap
pack does that out of the box.

Some test fixtures were killed along the way.
The code was moved to snapcraft.cli as most of the implementation left
in lifecycle.pack was mostly presentation related.

The downside here is that this makes it harder to support some commands
in Windows, but luckily Snapcraft for Windows is yet unreleased and the
broken commands (pack and upload-metadata) can eventually be
passthroughed through multipass.

Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>