Maven plugin: improve error message when target libs are not found.
It could just that maven-targets declaration does not match the
directory in which the Java library has been built. To help the
user better understand what could be wrong the error message has
been improved to add the directories in which the arfiles were
searched and clarifying a simple scenario (the one that happened
to me) that could lead to fix the situation.