build providers: check revision before switching
Snapcraft was switching channels on the availability of a tracking
channel, it may be that this was incorrectly implemented or that snapd
now switched to always report a tracking-channel when installing a snap
with --dangerous after previously having installed a revision from the
Snap Store.

An additional check for revision is now enforced before setting the
switch.

LP: #1884576
Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>