packaging: use PEP-440 compliant versioning for python & snap
If HEAD matches a tag, then version is simply the tag (e.g. 4.1.1).

If HEAD does not match a tag, then version indicating the distance to
last tag with the ".post" indicator and the git commit revision in
the local identifier (e.g. 4.1.1.post17+git92348a7eb).

Local identifiers should not be published publically, but since we only
push tagged versions to the pip index, this should not occur.

Now we can do package version comparisons, e.g.:
```
>>> from packaging import version
>>> version.parse("4.1.1") > version.parse("4.1.2")
False
>>> version.parse("4.1.1.post17+git92348a7eb") > version.parse("4.1.1")
True
```

Finally, match the snap version to the python package version to
ensure they stay in sync. Do this by querying setup.py for the
version number so only one implementation is required.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>