repo: install requested build-package versions
It either was never implemented, or has been broken for a long time.
However, it is documented as supported, so install the
build-package versions as requested.

This requires passing '--allow-downgrades' to apt, in case the user
changes the desired build-package.  Additionally, now the version
for each apt package is explicitly specified for apt-get install.

LP: #1887960

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>