extensions: use gtk3 platform theme on gtk-based DE's for kde (#3226)
The kde-frameworks-* platform snaps currently do not ship the gtk3 platform theme or QGnomePlatform. This is not an issue, however, because Qt will fallback to QGnomeTheme, which is a builtin theme and is always available.

Note: QGnomeTheme will use Xorg and XWayland to get theme info, so there are no additional connections required for snaps using it.