repo: consolidate BaseRepo and Ubuntu & reorganize package
There has been a long-running sync between abstract "BaseRepo" and
"Ubuntu".  The goal of the abstraction was to support multiple OSes,
but the reality of the snapcraft codebase today is that nearly every
usage of repo assumes Ubuntu (or at least debian/apt).

This drops the abstract "BaseRepo" and makes usage of repositories
explicitly tied to AptRepo.  Should the future support additional
repository types, we would need to do a lot more work to address
the split, as it's not as trivial as picking a repo per-platform.

- Drop DummyRepo.
- Drop repo._platform.
- Rename repo._deb to repo.apt.
- Create new repo.fixups module for the various fixup helpers,
  moving fix_pkg_config and _fix_file_mode into it.  More should
  be moved out of AptRepo in the future.
- Consolidate remaining repo._base utilities into repo.apt.
- Replace usage of "repo.Repo" with "repo.AptRepo"
- Replace usage of "repo.Ubuntu" with "repo.AptRepo"

This addresses import errors for "Ubuntu" on non-Ubuntu platforms,
as all users of AptRepo can safely import the repo package.  If
there was a bug and AptRepo were to be used on non-Ubuntu platforms,
it would result in an error from AptRepo being assigned None.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>